Now have bridge functions working, but no differential operator yet.
It remains a question whether the weights for the bridge functions should be
scaled. I think they probably should, such that under the quadrature rule with
the scaled weights, the bridge functions come out normalised.

It would be nice if papers talked about this and other things that a person
implementing this seems to just have to infer...